requirements.txt
streamlit>=1.33.0
ultralytics>=8.2.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0
numpy>=1.26.0
torch>=2.2.0
torchvision>=0.17.0
